PromptBot/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
attrs==24.2.0
azure-core==1.30.2
azure-monitor-opentelemetry-exporter==1.0.0b28
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
fastapi==0.112.0
filelock==3.15.4
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.2
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
MarkupSafe==2.1.5
marshmallow==3.21.3
more-itertools==10.4.0
msrest==0.7.1
numpy==2.0.1
oauthlib==3.2.2
openai==1.40.6
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
promptflow==1.15.0
promptflow-core==1.15.0
promptflow-devkit==1.15.0
promptflow-tracing==1.15.0
prompty==0.1.14
protobuf==4.25.4
psutil==5.9.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydash==7.0.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==2.1.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.20.0
